Move full height in a separate directive
