+++ client/app/scripts/liveblog-edit/directives.js
@@ -18,6 +18,7 @@ import selectTextOnClick from './directives/select-text-on-click';
 import lbBindHtml from './directives/bind-html';
 import lbFilterByMember from './directives/filter-by-member';
 import autofocus from './directives/autofocus';
+import fullHeight from './directives/full-height';
 
 angular.module('liveblog.edit')
     .directive('lbPostsList', lbPostsList)
@@ -164,43 +165,7 @@ angular.module('liveblog.edit')
     .directive('lbBindHtml', lbBindHtml)
     .directive('lbFilterByMember', lbFilterByMember)
     .directive('autofocus', autofocus)
-    //.directive('autofocus', ['$timeout', function($timeout) {
-    //    return {
-    //        restrict: 'A',
-    //        link: function($scope, $element) {
-    //            $timeout(function() {
-    //                $element[0].focus();
-    //            });
-    //        }
-    //    };
-    //}])
-    .directive('fullHeight', ['$timeout', '$window', 'lodash', function($timeout, $window, _) {
-        return {
-            restrict: 'A',
-            link: function($scope, $element, $attributes) {
-                // update the element height to the window height minus its vertical offset
-                function setHeight() {
-                    $timeout(function() {
-                        var height = $window.innerHeight - $element.offset().top;
-                        if ($attributes.fullHeightOffsetBottom) {
-                            height -= $attributes.fullHeightOffsetBottom;
-                        }
-                        var css_name = $attributes.fullHeightUseMaxHeight ? 'max-height' : 'height';
-                        $element.css(css_name, height);
-                        $element[0].focus();
-                    });
-                }
-                // initialize
-                setHeight();
-                // update when the window size changes
-                angular.element($window).on('resize', _.debounce(setHeight, 500));
-                // update when offset changes
-                $scope.$watch(function() {
-                    return $element.offset().top;
-                }, _.debounce(setHeight, 500));
-            }
-        };
-    }])
+    .directive('fullHeight', fullHeight)
     /**
     * Main directive to render the freetype editor.
     */
